# apps/server/api/openapi.yaml
openapi: 3.1.0
info:
  title: MMP v3 API
  version: 1.0.0
  description: Murder Mystery Platform v3 — Real-time multiplayer murder mystery game platform API.
  contact:
    name: MMP Team
servers:
  - url: /api/v1
security: []
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [Profile]
      summary: Get own profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateProfile
      tags: [Profile]
      summary: Update own profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{id}:
    get:
      operationId: getPublicProfile
      tags: [Profile]
      summary: Get public profile by user ID
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfileResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Themes (public) ────────────────────────────────────
...
  /themes/{id}:
    get:
      operationId: getTheme
      tags: [Themes]
      summary: Get theme by ID
      parameters:
        - $ref: "#/components/parameters/ThemeIdPath"
      responses:
        "200":
          description: Theme detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /themes/slug/{slug}:
    get:
      operationId: getThemeBySlug
      tags: [Themes]
      summary: Get theme by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Theme detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /rooms:
    get:
      operationId: listRooms
      tags: [Rooms]
      summary: List public rooms
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Room list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomResponse"
    post:
      operationId: createRoom
      tags: [Rooms]
      summary: Create a new room
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [Rooms]
      summary: Get room detail with players
      parameters:
        - $ref: "#/components/parameters/RoomIdPath"
      responses:
        "200":
          description: Room detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /rooms/code/{code}:
    get:
      operationId: getRoomByCode
      tags: [Rooms]
      summary: Get room by invite code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Room detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /rooms/{id}/join:
    post:
      operationId: joinRoom
      tags: [Rooms]
      summary: Join a room
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomIdPath"
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /rooms/{id}/leave:
    post:
      operationId: leaveRoom
      tags: [Rooms]
      summary: Leave a room
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomIdPath"
      responses:
        "200":
          description: Left the room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Editor ─────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Parameters ──────────────────────────────────────────
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    UserIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ThemeIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RoomIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Conflict:
      description: Conflict with current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"

  # ─── Schemas ─────────────────────────────────────────────
  schemas:
...
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          description: URI reference identifying the problem type
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation
        code:
          type: string
          description: Application-specific error code
        instance:
          type: string
          description: URI reference identifying the specific occurrence

    # ── Generic ────────────────────────────────────────────
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ───────────────────────────────────────────────
...
    UpdateProfileRequest:
      type: object
      required: [nickname]
      properties:
        nickname:
          type: string
          minLength: 2
          maxLength: 30
        avatar_url:
          type: string
          format: uri
    ProfileResponse:
      type: object
      required: [id, nickname, role, coin_balance]
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        email:
          type: string
          format: email
        avatar_url:
          type: string
          format: uri
        role:
          type: string
        coin_balance:
          type: integer
    PublicProfileResponse:
      type: object
      required: [id, nickname]
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        avatar_url:
          type: string
          format: uri

    # ── Theme (public) ─────────────────────────────────────
...
    ThemeResponse:
      type: object
      required: [id, title, slug, min_players, max_players, duration_min, price, creator_id, status, version, created_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        cover_image:
          type: string
          format: uri
        min_players:
          type: integer
        max_players:
          type: integer
        duration_min:
          type: integer
        price:
          type: integer
        creator_id:
          type: string
          format: uuid
        status:
          type: string
        config_json:
          type: object
        version:
          type: integer
        published_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    CreateRoomRequest:
      type: object
      required: [theme_id, max_players, is_private]
      properties:
        theme_id:
          type: string
          format: uuid
        max_players:
          type: integer
          minimum: 2
          maximum: 20
        is_private:
          type: boolean
    RoomResponse:
      type: object
      required: [id, theme_id, host_id, code, status, max_players, is_private, player_count, created_at]
      properties:
        id:
          type: string
          format: uuid
        theme_id:
          type: string
          format: uuid
        host_id:
          type: string
          format: uuid
        code:
          type: string
        status:
          type: string
        max_players:
          type: integer
        is_private:
          type: boolean
        player_count:
          type: integer
        created_at:
          type: string
          format: date-time
    PlayerInfo:
      type: object
      required: [user_id, is_ready]
      properties:
        user_id:
          type: string
          format: uuid
        is_ready:
          type: boolean
    RoomDetailResponse:
      type: object
      required: [id, theme_id, host_id, code, status, max_players, is_private, player_count, created_at, players]
      properties:
        id:
          type: string
          format: uuid
        theme_id:
          type: string
          format: uuid
        host_id:
          type: string
          format: uuid
        code:
          type: string
        status:
          type: string
        max_players:
          type: integer
        is_private:
          type: boolean
        player_count:
          type: integer
        created_at:
          type: string
          format: date-time
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerInfo"

    # ── Editor ─────────────────────────────────────────────
...